Simplify operator event handling
Add 'operators' dictionary that maps operator strings (i.e. '+', '-',
'*', '/') to real mathematic operators using the 'operator' module.

Remove calculator.py and reference to its functions in
calculator-gui.py, as these functions are redundant.

Change name of 'operator' variable to 'current_operator', so that it
does not override the imported 'operator' module, and more accurately
reflect its purpose - to hold the string of the operator to apply in the
current expression.